+++ portal-impl/src/com/liferay/portal/service/http/OrganizationServiceHttp.java
@@ -85,8 +85,8 @@ public class OrganizationServiceHttp {
 	public static com.liferay.portal.model.Organization addOrganization(
 		HttpPrincipal httpPrincipal, long parentOrganizationId,
 		java.lang.String name, java.lang.String type, boolean recursable,
-		long regionId, long countryId, int statusId, java.lang.String comments,
-		boolean site,
+		long regionId, long countryId, long statusId,
+		java.lang.String comments, boolean site,
 		java.util.List<com.liferay.portal.model.Address> addresses,
 		java.util.List<com.liferay.portal.model.EmailAddress> emailAddresses,
 		java.util.List<com.liferay.portal.model.OrgLabor> orgLabors,
@@ -128,8 +128,9 @@ public class OrganizationServiceHttp {
 	public static com.liferay.portal.model.Organization addOrganization(
 		HttpPrincipal httpPrincipal, long parentOrganizationId,
 		java.lang.String name, java.lang.String type, boolean recursable,
-		long regionId, long countryId, int statusId, java.lang.String comments,
-		boolean site, com.liferay.portal.service.ServiceContext serviceContext)
+		long regionId, long countryId, long statusId,
+		java.lang.String comments, boolean site,
+		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		try {
 			MethodKey methodKey = new MethodKey(OrganizationServiceUtil.class,
@@ -164,7 +165,7 @@ public class OrganizationServiceHttp {
 	public static com.liferay.portal.model.Organization addOrganization(
 		HttpPrincipal httpPrincipal, long parentOrganizationId,
 		java.lang.String name, java.lang.String type, long regionId,
-		long countryId, int statusId, java.lang.String comments, boolean site,
+		long countryId, long statusId, java.lang.String comments, boolean site,
 		java.util.List<com.liferay.portal.model.Address> addresses,
 		java.util.List<com.liferay.portal.model.EmailAddress> emailAddresses,
 		java.util.List<com.liferay.portal.model.OrgLabor> orgLabors,
@@ -206,7 +207,7 @@ public class OrganizationServiceHttp {
 	public static com.liferay.portal.model.Organization addOrganization(
 		HttpPrincipal httpPrincipal, long parentOrganizationId,
 		java.lang.String name, java.lang.String type, long regionId,
-		long countryId, int statusId, java.lang.String comments, boolean site,
+		long countryId, long statusId, java.lang.String comments, boolean site,
 		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		try {
@@ -629,7 +630,7 @@ public class OrganizationServiceHttp {
 		HttpPrincipal httpPrincipal, long organizationId,
 		long parentOrganizationId, java.lang.String name,
 		java.lang.String type, boolean recursable, long regionId,
-		long countryId, int statusId, java.lang.String comments, boolean site,
+		long countryId, long statusId, java.lang.String comments, boolean site,
 		java.util.List<com.liferay.portal.model.Address> addresses,
 		java.util.List<com.liferay.portal.model.EmailAddress> emailAddresses,
 		java.util.List<com.liferay.portal.model.OrgLabor> orgLabors,
@@ -673,7 +674,7 @@ public class OrganizationServiceHttp {
 		HttpPrincipal httpPrincipal, long organizationId,
 		long parentOrganizationId, java.lang.String name,
 		java.lang.String type, boolean recursable, long regionId,
-		long countryId, int statusId, java.lang.String comments, boolean site,
+		long countryId, long statusId, java.lang.String comments, boolean site,
 		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		try {
@@ -710,7 +711,7 @@ public class OrganizationServiceHttp {
 	public static com.liferay.portal.model.Organization updateOrganization(
 		HttpPrincipal httpPrincipal, long organizationId,
 		long parentOrganizationId, java.lang.String name,
-		java.lang.String type, long regionId, long countryId, int statusId,
+		java.lang.String type, long regionId, long countryId, long statusId,
 		java.lang.String comments, boolean logo, byte[] logoBytes,
 		boolean site,
 		java.util.List<com.liferay.portal.model.Address> addresses,
@@ -755,7 +756,7 @@ public class OrganizationServiceHttp {
 	public static com.liferay.portal.model.Organization updateOrganization(
 		HttpPrincipal httpPrincipal, long organizationId,
 		long parentOrganizationId, java.lang.String name,
-		java.lang.String type, long regionId, long countryId, int statusId,
+		java.lang.String type, long regionId, long countryId, long statusId,
 		java.lang.String comments, boolean site,
 		java.util.List<com.liferay.portal.model.Address> addresses,
 		java.util.List<com.liferay.portal.model.EmailAddress> emailAddresses,
@@ -798,7 +799,7 @@ public class OrganizationServiceHttp {
 	public static com.liferay.portal.model.Organization updateOrganization(
 		HttpPrincipal httpPrincipal, long organizationId,
 		long parentOrganizationId, java.lang.String name,
-		java.lang.String type, long regionId, long countryId, int statusId,
+		java.lang.String type, long regionId, long countryId, long statusId,
 		java.lang.String comments, boolean site,
 		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
@@ -838,7 +839,7 @@ public class OrganizationServiceHttp {
 		};
 	private static final Class<?>[] _addOrganizationParameterTypes1 = new Class[] {
 			long.class, java.lang.String.class, java.lang.String.class,
-			boolean.class, long.class, long.class, int.class,
+			boolean.class, long.class, long.class, long.class,
 			java.lang.String.class, boolean.class, java.util.List.class,
 			java.util.List.class, java.util.List.class, java.util.List.class,
 			java.util.List.class,
@@ -846,20 +847,20 @@ public class OrganizationServiceHttp {
 		};
 	private static final Class<?>[] _addOrganizationParameterTypes2 = new Class[] {
 			long.class, java.lang.String.class, java.lang.String.class,
-			boolean.class, long.class, long.class, int.class,
+			boolean.class, long.class, long.class, long.class,
 			java.lang.String.class, boolean.class,
 			com.liferay.portal.service.ServiceContext.class
 		};
 	private static final Class<?>[] _addOrganizationParameterTypes3 = new Class[] {
 			long.class, java.lang.String.class, java.lang.String.class,
-			long.class, long.class, int.class, java.lang.String.class,
+			long.class, long.class, long.class, java.lang.String.class,
 			boolean.class, java.util.List.class, java.util.List.class,
 			java.util.List.class, java.util.List.class, java.util.List.class,
 			com.liferay.portal.service.ServiceContext.class
 		};
 	private static final Class<?>[] _addOrganizationParameterTypes4 = new Class[] {
 			long.class, java.lang.String.class, java.lang.String.class,
-			long.class, long.class, int.class, java.lang.String.class,
+			long.class, long.class, long.class, java.lang.String.class,
 			boolean.class, com.liferay.portal.service.ServiceContext.class
 		};
 	private static final Class<?>[] _addPasswordPolicyOrganizationsParameterTypes5 =
@@ -902,7 +903,7 @@ public class OrganizationServiceHttp {
 	private static final Class<?>[] _updateOrganizationParameterTypes18 = new Class[] {
 			long.class, long.class, java.lang.String.class,
 			java.lang.String.class, boolean.class, long.class, long.class,
-			int.class, java.lang.String.class, boolean.class,
+			long.class, java.lang.String.class, boolean.class,
 			java.util.List.class, java.util.List.class, java.util.List.class,
 			java.util.List.class, java.util.List.class,
 			com.liferay.portal.service.ServiceContext.class
@@ -910,12 +911,12 @@ public class OrganizationServiceHttp {
 	private static final Class<?>[] _updateOrganizationParameterTypes19 = new Class[] {
 			long.class, long.class, java.lang.String.class,
 			java.lang.String.class, boolean.class, long.class, long.class,
-			int.class, java.lang.String.class, boolean.class,
+			long.class, java.lang.String.class, boolean.class,
 			com.liferay.portal.service.ServiceContext.class
 		};
 	private static final Class<?>[] _updateOrganizationParameterTypes20 = new Class[] {
 			long.class, long.class, java.lang.String.class,
-			java.lang.String.class, long.class, long.class, int.class,
+			java.lang.String.class, long.class, long.class, long.class,
 			java.lang.String.class, boolean.class, byte[].class, boolean.class,
 			java.util.List.class, java.util.List.class, java.util.List.class,
 			java.util.List.class, java.util.List.class,
@@ -923,7 +924,7 @@ public class OrganizationServiceHttp {
 		};
 	private static final Class<?>[] _updateOrganizationParameterTypes21 = new Class[] {
 			long.class, long.class, java.lang.String.class,
-			java.lang.String.class, long.class, long.class, int.class,
+			java.lang.String.class, long.class, long.class, long.class,
 			java.lang.String.class, boolean.class, java.util.List.class,
 			java.util.List.class, java.util.List.class, java.util.List.class,
 			java.util.List.class,
@@ -931,7 +932,7 @@ public class OrganizationServiceHttp {
 		};
 	private static final Class<?>[] _updateOrganizationParameterTypes22 = new Class[] {
 			long.class, long.class, java.lang.String.class,
-			java.lang.String.class, long.class, long.class, int.class,
+			java.lang.String.class, long.class, long.class, long.class,
 			java.lang.String.class, boolean.class,
 			com.liferay.portal.service.ServiceContext.class
 		};
